refactor: streamline QR scanning and uploading process
- Replaced onScanSuccess with onFinish in ScanQrView to unify callback handling.
- Introduced QrScanner component to encapsulate QR scanning logic, improving code organization.
- Updated UploadQrPrompt to utilize useCoreNavigate for navigation, enhancing user experience.
- Refactored UploadQrPage components to align with new scanning and uploading structure.
